requirements.txt
coverage==5.5
docker==5.0.2
pydantic==1.8.2
pytest==6.2.5
PyYAML==5.4.1
scylla-driver==3.24.5
